requirements.txt
discord.py==2.3.2
google-generativeai==0.3.2
Pillow==9.5.0
requests==2.28.2
flask==2.3.3